+++ config/inertia.php
@@ -10,6 +10,9 @@ return [
     | These options configures if and how Inertia uses Server Side Rendering
     | to pre-render the initial visits made to your application's pages.
     |
+    | You can specify a custom SSR bundle path, or omit it to let Inertia
+    | try and automatically detect it for you.
+    |
     | Do note that enabling these options will NOT automatically make SSR work,
     | as a separate rendering service needs to be available. To learn more,
     | please visit https://inertiajs.com/server-side-rendering
@@ -18,9 +21,11 @@ return [
 
     'ssr' => [
 
-        'enabled' => false,
+        'enabled' => true,
 
-        'url' => 'http://127.0.0.1:13714/render',
+        'url' => 'http://127.0.0.1:13714',
+
+        // 'bundle' => base_path('bootstrap/ssr/ssr.mjs'),
 
     ],