+++ src/Directive.php
@@ -7,20 +7,20 @@ class Directive
     /**
      * Compiles the "@inertia" directive.
      *
-     * @param  string  $expression
-     * @return string
+     * @param string $expression
      */
     public static function compile($expression = ''): string
     {
         $id = trim(trim($expression), "\'\"") ?: 'app';
 
         $template = '<?php
-            if (!isset($__inertiaSsr)) {
-                $__inertiaSsr = app(\Inertia\Ssr\Gateway::class)->dispatch($page);
+            if (!isset($__inertiaSsrDispatched)) {
+                $__inertiaSsrDispatched = true;
+                $__inertiaSsrResponse = app(\Inertia\Ssr\Gateway::class)->dispatch($page);
             }
 
-            if ($__inertiaSsr instanceof \Inertia\Ssr\Response) {
-                echo $__inertiaSsr->body;
+            if ($__inertiaSsrResponse) {
+                echo $__inertiaSsrResponse->body;
             } else {
                 ?><div id="'.$id.'" data-page="{{ json_encode($page) }}"></div><?php
             }
@@ -32,18 +32,18 @@ class Directive
     /**
      * Compiles the "@inertiaHead" directive.
      *
-     * @param  string  $expression
-     * @return string
+     * @param string $expression
      */
     public static function compileHead($expression = ''): string
     {
         $template = '<?php
-            if (!isset($__inertiaSsr)) {
-                $__inertiaSsr = app(\Inertia\Ssr\Gateway::class)->dispatch($page);
+            if (!isset($__inertiaSsrDispatched)) {
+                $__inertiaSsrDispatched = true;
+                $__inertiaSsrResponse = app(\Inertia\Ssr\Gateway::class)->dispatch($page);
             }
 
-            if ($__inertiaSsr instanceof \Inertia\Ssr\Response) {
-                echo $__inertiaSsr->head;
+            if ($__inertiaSsrResponse) {
+                echo $__inertiaSsrResponse->head;
             }
         ?>';