+++ src/ResponseFactory.php
@@ -3,13 +3,15 @@
 namespace Inertia;
 
 use Closure;
-use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Request;
-use Illuminate\Support\Facades\Response as BaseResponse;
 use Illuminate\Support\Traits\Macroable;
+use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Response as BaseResponse;
+use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse as SymfonyRedirect;
 
 class ResponseFactory
 {
@@ -30,8 +32,8 @@ class ResponseFactory
     }
 
     /**
-     * @param  string|array|Arrayable  $key
-     * @param  mixed|null  $value
+     * @param string|array|Arrayable $key
+     * @param mixed                  $value
      */
     public function share($key, $value = null): void
     {
@@ -45,8 +47,8 @@ class ResponseFactory
     }
 
     /**
-     * @param  string|null  $key
-     * @param  null|mixed  $default
+     * @param mixed $default
+     *
      * @return mixed
      */
     public function getShared(string $key = null, $default = null)
@@ -64,7 +66,7 @@ class ResponseFactory
     }
 
     /**
-     * @param  Closure|string|null  $version
+     * @param Closure|string|null $version
      */
     public function version($version): void
     {
@@ -86,9 +88,7 @@ class ResponseFactory
     }
 
     /**
-     * @param  string  $component
-     * @param  array|Arrayable  $props
-     * @return Response
+     * @param array|Arrayable $props
      */
     public function render(string $component, $props = []): Response
     {
@@ -105,18 +105,14 @@ class ResponseFactory
     }
 
     /**
-     * @param  string|RedirectResponse  $url
+     * @param string|SymfonyRedirect $url
      */
-    public function location($url): \Symfony\Component\HttpFoundation\Response
+    public function location($url): SymfonyResponse
     {
-        if ($url instanceof RedirectResponse) {
-            $url = $url->getTargetUrl();
-        }
-
         if (Request::inertia()) {
-            return BaseResponse::make('', 409, ['X-Inertia-Location' => $url]);
+            return BaseResponse::make('', 409, ['X-Inertia-Location' => $url instanceof SymfonyRedirect ? $url->getTargetUrl() : $url]);
         }
 
-        return new RedirectResponse($url);
+        return $url instanceof SymfonyRedirect ? $url : Redirect::away($url);
     }
 }