+++ src/Ssr/HttpGateway.php
@@ -3,24 +3,20 @@
 namespace Inertia\Ssr;
 
 use Exception;
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Http;
 
 class HttpGateway implements Gateway
 {
     /**
      * Dispatch the Inertia page to the Server Side Rendering engine.
-     *
-     * @param  array  $page
-     * @return Response|null
      */
     public function dispatch(array $page): ?Response
     {
-        if (! Config::get('inertia.ssr.enabled', false)) {
+        if (! config('inertia.ssr.enabled', true) || ! (new BundleDetector())->detect()) {
             return null;
         }
 
-        $url = Config::get('inertia.ssr.url', 'http://127.0.0.1:13714/render');
+        $url = str_replace('/render', '', config('inertia.ssr.url', 'http://127.0.0.1:13714')).'/render';
 
         try {
             $response = Http::post($url, $page)->throw()->json();