+++ src/Ssr/SsrException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Inertia\Ssr;
+
+use Exception;
+
+class SsrException extends Exception
+{
+}