+++ tests/Testing/TestResponseMacrosTest.php
@@ -2,10 +2,9 @@
 
 namespace Inertia\Tests\Testing;
 
-use Illuminate\Testing\Fluent\AssertableJson;
 use Inertia\Inertia;
-use Inertia\Testing\Assert;
 use Inertia\Tests\TestCase;
+use Illuminate\Testing\Fluent\AssertableJson;
 
 class TestResponseMacrosTest extends TestCase
 {
@@ -18,12 +17,7 @@ class TestResponseMacrosTest extends TestCase
 
         $success = false;
         $response->assertInertia(function ($page) use (&$success) {
-            if (class_exists(AssertableJson::class)) {
-                $this->assertInstanceOf(AssertableJson::class, $page);
-            } else {
-                // TODO: Remove once built-in Assert library is removed.
-                $this->assertInstanceOf(Assert::class, $page);
-            }
+            $this->assertInstanceOf(AssertableJson::class, $page);
             $success = true;
         });